# desktop/src/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "CacheClearResponse": {
        "properties": {
          "removed": {
            "title": "Removed",
            "type": "integer"
          },
          "scope": {
            "enum": [
              "expired",
              "all"
            ],
            "title": "Scope",
            "type": "string"
          },
          "target": {
            "enum": [
              "paragraph",
              "pdf",
              "all"
            ],
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "removed",
          "scope",
          "target"
        ],
        "title": "CacheClearResponse",
        "type": "object"
      },
      "CacheOverviewResponse": {
        "description": "Both translation caches, which Settings \u2192 Cache shows side by side. It\nused to show only the paragraph cache, while its \"Clear all\" emptied both\n(#32).",
        "properties": {
          "paragraph": {
            "$ref": "#/components/schemas/CacheStatsResponse"
          },
          "pdf": {
            "$ref": "#/components/schemas/PdfCacheStatsResponse"
          }
        },
        "required": [
          "paragraph",
          "pdf"
        ],
        "title": "CacheOverviewResponse",
        "type": "object"
      },
...
      "ConfigResponse": {
        "properties": {
          "anthropic": {
            "$ref": "#/components/schemas/APIKeyMaskedSettings"
          },
          "argos": {
            "$ref": "#/components/schemas/APIKeyMaskedSettings"
          },
          "debug_mode": {
            "title": "Debug Mode",
            "type": "boolean"
          },
          "gemini": {
            "$ref": "#/components/schemas/APIKeyMaskedSettings"
          },
          "gui": {
            "$ref": "#/components/schemas/GUISettings"
          },
          "openai": {
            "$ref": "#/components/schemas/APIKeyMaskedSettings"
          },
          "processing": {
            "$ref": "#/components/schemas/ProcessingSettings"
          },
          "rag": {
            "$ref": "#/components/schemas/RAGSettings"
          },
          "translation": {
            "$ref": "#/components/schemas/TranslationConfig"
          }
        },
        "required": [
          "openai",
          "gemini",
          "anthropic",
          "argos",
          "translation",
          "rag",
          "gui",
          "processing",
          "debug_mode"
        ],
        "title": "ConfigResponse",
        "type": "object"
      },
      "ConfigUpdateRequest": {
        "properties": {
          "answer_model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "anthropic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EndpointCredentialUpdate"
              },
              {
                "type": "null"
              }
            ]
          },
          "cache_translated_pdfs": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Translated Pdfs"
          },
          "cache_translations": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Translations"
          },
          "chat_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Enabled"
          },
          "default_source_lang": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LanguageCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "default_target_lang": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LanguageCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "gemini": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceCredentialUpdate"
              },
              {
                "type": "null"
              }
            ]
          },
          "max_file_size_mb": {
            "anyOf": [
              {
                "maximum": 200.0,
                "minimum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max File Size Mb"
          },
          "max_pages": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Pages"
          },
          "max_parallel_chunks": {
            "anyOf": [
              {
                "maximum": 16.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Parallel Chunks"
          },
          "openai": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EndpointCredentialUpdate"
              },
              {
                "type": "null"
              }
            ]
          },
          "preferred_service": {
            "anyOf": [
              {
                "description": "A provider's id, as GET /providers lists them",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Service"
          },
          "translation_model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ConfigUpdateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "PDFusion sidecar",
    "version": "1.2.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/config": {
      "get": {
        "operationId": "get_config_config_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Config",
        "tags": [
          "config"
        ]
      },
      "put": {
        "operationId": "update_config_config_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Config",
        "tags": [
          "config"
        ]
      }
    },
    "/config/cache": {
      "delete": {
        "description": "Clear the on-disk translation caches.\n\n`target` names which: `paragraph` (the default), `pdf` (the whole-PDF\ncache) or `all`. `scope=expired` reaps only entries past their TTL, which\nonly the paragraph cache has, so it never touches the PDF cache. Anything\nelse is a 422: a mistyped target used to fall through to clearing.",
        "operationId": "clear_cache_config_cache_delete",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "all",
                "expired"
              ],
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "default": "paragraph",
              "enum": [
                "paragraph",
                "pdf",
                "all"
              ],
              "title": "Target",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheClearResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Cache",
        "tags": [
          "config"
        ]
      },
      "get": {
        "operationId": "get_cache_stats_config_cache_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cache Stats",
        "tags": [
          "config"
        ]
      }
    },
    "/config/models/{service}": {
      "get": {
        "description": "The ids the saved key can use at the saved endpoint.\n\nA wrapper over the catalog (`GET /providers/{id}/models`) for the\ntoolbar's model picker: listed models only \u2014 the picker adds the saved\nmodel and, without a list, the suggestions itself. Always the saved key\nwith the saved endpoint, the pair `PUT /config` keeps together, so this\ncan't send a key anywhere new.",
        "operationId": "list_endpoint_models_config_models__service__get",
        "parameters": [
          {
            "in": "path",
            "name": "service",
            "required": true,
            "schema": {
              "description": "A provider's id, as GET /providers lists them",
              "title": "Service",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointModelsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Endpoint Models",
        "tags": [
          "config"
        ]
      }
    },
    "/config/options": {
      "get": {
        "operationId": "get_options_config_options_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Options",
        "tags": [
          "config"
        ]
      }
    },
    "/config/validate": {
      "post": {
        "description": "Check credentials with the provider, by listing the key's models.\n\nA wrapper over `POST /providers/{id}/verify`. Whatever the request leaves\nout comes from the saved settings: the key, the model, the endpoint. The\nsaved key is only sent to the saved endpoint, the rule `PUT /config`\nkeeps; checking another endpoint needs the key typed alongside it.",
        "operationId": "validate_credentials_config_validate_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Credentials",
        "tags": [
          "config"
        ]
      }
    },
...
    "/providers/{provider_id}/models": {
      "get": {
        "operationId": "provider_models_providers__provider_id__models_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "description": "A provider's id, as GET /providers lists them",
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "title": "Refresh",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provider Models",
        "tags": [
          "providers"
        ]
      }
    },
    "/providers/{provider_id}/verify": {
      "post": {
        "operationId": "verify_provider_providers__provider_id__verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "description": "A provider's id, as GET /providers lists them",
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Provider",
        "tags": [
          "providers"
        ]
      }
    },
...
  }
}